avoid remove_roster/2 as it crashes
I realise that doing a full wipe + re-add of users is overkill, but it is
good enough for now, and lets me get on with my life.

For reference, the error exists for
[add_contact, subscribe, subscribe_decline, unsubscribe]
and is printed as:

=== WARNING: end_per_testcase crashed!
Reason: badarg
Line: [{lists,member},
 {presence_SUITE,remove_roster,385},
 {presence_SUITE,end_per_testcase,68},
 {test_server,do_end_per_testcase,1655},
 {test_server,run_test_case_eval1,1392},
 {test_server,run_test_case_eval,1302}]

It causes the second test in subscribe_group to fail,
and the rest to be skipped.